+++ src/publishers/UbidotsPublisher.cpp
@@ -32,27 +32,30 @@ const char* UbidotsPublisher::payload = "{";
 
 
 // Constructors
-UbidotsPublisher::UbidotsPublisher() : dataPublisher() {}
-UbidotsPublisher::UbidotsPublisher(Logger& baseLogger, int sendEveryX)
-    : dataPublisher(baseLogger, sendEveryX) {}
-UbidotsPublisher::UbidotsPublisher(Logger& baseLogger, Client* inClient,
-                                   int sendEveryX)
-    : dataPublisher(baseLogger, inClient, sendEveryX) {}
+// Primary constructor with authentication parameters
 UbidotsPublisher::UbidotsPublisher(Logger&     baseLogger,
                                    const char* authenticationToken,
                                    const char* deviceID, int sendEveryX)
     : dataPublisher(baseLogger, sendEveryX) {
-    setToken(authenticationToken);
-    _baseLogger->setSamplingFeatureUUID(deviceID);
+    if (authenticationToken) setToken(authenticationToken);
+    if (deviceID) _baseLogger->setSamplingFeatureUUID(deviceID);
     MS_DBG(F("dataPublisher object created"));
+}
+
+// Delegating constructors
+UbidotsPublisher::UbidotsPublisher() : dataPublisher() {}
+UbidotsPublisher::UbidotsPublisher(Logger& baseLogger, int sendEveryX)
+    : UbidotsPublisher(baseLogger, nullptr, nullptr, sendEveryX) {}
+UbidotsPublisher::UbidotsPublisher(Logger& baseLogger, Client* inClient,
+                                   int sendEveryX)
+    : UbidotsPublisher(baseLogger, nullptr, nullptr, sendEveryX) {
+    if (inClient) _inClient = inClient;
 }
 UbidotsPublisher::UbidotsPublisher(Logger& baseLogger, Client* inClient,
                                    const char* authenticationToken,
                                    const char* deviceID, int sendEveryX)
-    : dataPublisher(baseLogger, inClient, sendEveryX) {
-    setToken(authenticationToken);
-    _baseLogger->setSamplingFeatureUUID(deviceID);
-    MS_DBG(F("dataPublisher object created"));
+    : UbidotsPublisher(baseLogger, authenticationToken, deviceID, sendEveryX) {
+    if (inClient) _inClient = inClient;
 }
 // Destructor
 UbidotsPublisher::~UbidotsPublisher() {}