+++ src/sensors/AtlasParent.cpp
@@ -25,16 +25,15 @@ AtlasParent::AtlasParent(TwoWire* theI2C, int8_t powerPin,
              measurementsToAverage, incCalcValues),
       _i2cAddressHex(i2cAddressHex),
       _i2c(theI2C) {}
+// Delegating constructor
 AtlasParent::AtlasParent(int8_t powerPin, uint8_t i2cAddressHex,
                          uint8_t measurementsToAverage, const char* sensorName,
                          const uint8_t totalReturnedValues,
                          uint32_t warmUpTime_ms, uint32_t stabilizationTime_ms,
                          uint32_t measurementTime_ms, uint8_t incCalcValues)
-    : Sensor(sensorName, totalReturnedValues, warmUpTime_ms,
-             stabilizationTime_ms, measurementTime_ms, powerPin, -1,
-             measurementsToAverage, incCalcValues),
-      _i2cAddressHex(i2cAddressHex),
-      _i2c(&Wire) {}
+    : AtlasParent(&Wire, powerPin, i2cAddressHex, measurementsToAverage,
+                  sensorName, totalReturnedValues, warmUpTime_ms,
+                  stabilizationTime_ms, measurementTime_ms, incCalcValues) {}
 // Destructors
 AtlasParent::~AtlasParent() {}