+++ src/sensors/KellerParent.cpp
@@ -29,6 +29,7 @@ KellerParent::KellerParent(byte modbusAddress, Stream* stream, int8_t powerPin,
       _RS485EnablePin(enablePin) {
     setSecondaryPowerPin(powerPin2);
 }
+// Delegating constructor
 KellerParent::KellerParent(byte modbusAddress, Stream& stream, int8_t powerPin,
                            int8_t powerPin2, int8_t enablePin,
                            uint8_t measurementsToAverage, kellerModel model,
@@ -36,15 +37,9 @@ KellerParent::KellerParent(byte modbusAddress, Stream& stream, int8_t powerPin,
                            uint32_t warmUpTime_ms,
                            uint32_t stabilizationTime_ms,
                            uint32_t measurementTime_ms)
-    : Sensor(sensName, numVariables, warmUpTime_ms, stabilizationTime_ms,
-             measurementTime_ms, powerPin, -1, measurementsToAverage,
-             KELLER_INC_CALC_VARIABLES),
-      _model(model),
-      _modbusAddress(modbusAddress),
-      _stream(&stream),
-      _RS485EnablePin(enablePin) {
-    setSecondaryPowerPin(powerPin2);
-}
+    : KellerParent(modbusAddress, &stream, powerPin, powerPin2, enablePin,
+                   measurementsToAverage, model, sensName, numVariables,
+                   warmUpTime_ms, stabilizationTime_ms, measurementTime_ms) {}
 // Destructor
 KellerParent::~KellerParent() {}